Get it working
This is actually able to successfully extract the gpio start and end
addresses on its own.

Big thing is to ensure a large K so that similarity search finds as much
as possible that is relevant, then filter out by the similarity score.

Then I convert the documents to YAML.

Using Haiku 4.5 seems to be pretty good since it is quite fast and from
what I've seen it isn't too far from sonnet in skill

Surprised this works without special handling for tabular information..
yet.

- Adds a bunch of new model ids for new models. Shouldn't be using haiku
  3 for sure.

- Adds a generic documents_to_yaml function